requirements.txt
flask
gunicorn
binance-connector
pyTelegramBotAPI